refactor(hook): extract shared external-write body module
Two hooks carried semantically identical copies of external-write surface
detection and body extraction. Both specs recorded the repo 2-copy rule with
the same directive to extract at the third consumer; caller-probe-gate (#906)
is that consumer.

Compared all 13 shared functions and constants before moving: differences
were comment wording only, so the migration is behavior-preserving and both
existing suites (34 + 60 cases) act as the parity oracle.

Aliased to the previous private names at each import site so call sites and
tests stay untouched.

Confidence: high
Scope-risk: narrow
